@@ -1,11 +1,11 @@
 class Country {
-  final String name;
+  final String? name;
 
-  final String flagUri;
+  final String? flagUri;
 
-  final String code;
+  final String? code;
 
-  final String dialCode;
+  final String? dialCode;
 
   Country({this.name, this.code, this.flagUri, this.dialCode});
 }
