@@ -7,13 +7,13 @@ import 'package:flutter/services.dart';
 
 class PhoneService {
   static List<Country> getPotentialCountries(
-      String number, List<Country> countries) {
+      String number, List<Country>? countries) {
     List<Country> result = [];
     if (number.length > 0 && number.length < 5) {
       List<String> potentialCodes =
           generatePotentialDialCodes(number, 0, number.length);
       for (var code in potentialCodes) {
-        for (var country in countries) {
+        for (var country in countries!) {
           if (code == country.dialCode) {
             result.add(country);
           }
@@ -24,7 +24,7 @@ class PhoneService {
       String intlCode = number.substring(0, 4);
       List<String> potentialCodes = generatePotentialDialCodes(intlCode, 0, 4);
       for (var code in potentialCodes) {
-        for (var country in countries) {
+        for (var country in countries!) {
           if (code == country.dialCode) {
             result.add(country);
           }
@@ -48,9 +48,9 @@ class PhoneService {
     return potentialCodes;
   }
 
-  static Future<bool> parsePhoneNumber(String number, String iso) async {
+  static Future<bool?> parsePhoneNumber(String number, String iso) async {
     try {
-      bool isValid = await PhoneNumberUtil.isValidPhoneNumber(
+      bool? isValid = await PhoneNumberUtil.isValidPhoneNumber(
           phoneNumber: number, isoCode: iso);
       return isValid;
     } on PlatformException {
@@ -58,10 +58,10 @@ class PhoneService {
     }
   }
 
-  static Future<String> getNormalizedPhoneNumber(
+  static Future<String?> getNormalizedPhoneNumber(
       String number, String iso) async {
     try {
-      String normalizedNumber = await PhoneNumberUtil.normalizePhoneNumber(
+      String? normalizedNumber = await PhoneNumberUtil.normalizePhoneNumber(
           phoneNumber: number, isoCode: iso);
 
       return normalizedNumber;
